pyproject.toml
[project]
name = "cli-apollo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-cpp-python>=0.3.14",
    "rich>=14.0.0",
]
